+++ src/management/gpu/archs/dummy.h
@@ -37,14 +37,14 @@ state_t mgt_gpu_dummy_freq_limit_set(ctx_t *c, ulong *khz);
 
 state_t mgt_gpu_dummy_freq_get_available(ctx_t *c, const ulong ***list_khz, const uint **list_len);
 
-state_t mgt_gpu_dummy_power_cap_get_current(ctx_t *c, ulong *watts);
+state_t mgt_gpu_dummy_power_cap_get_current(ctx_t *c, uint32_t *watts);
 
-state_t mgt_gpu_dummy_power_cap_get_default(ctx_t *c, ulong *watts);
+state_t mgt_gpu_dummy_power_cap_get_default(ctx_t *c, uint32_t *watts);
 
-state_t mgt_gpu_dummy_power_cap_get_rank(ctx_t *c, ulong *watts_min, ulong *watts_max);
+state_t mgt_gpu_dummy_power_cap_get_rank(ctx_t *c, uint32_t *watts_min, uint32_t *watts_max);
 
 state_t mgt_gpu_dummy_power_cap_reset(ctx_t *c);
 
-state_t mgt_gpu_dummy_power_cap_set(ctx_t *c, ulong *watts);
+state_t mgt_gpu_dummy_power_cap_set(ctx_t *c, uint32_t *watts);
 
 #endif
